Fix code benchmark paths and namespaces.
