+++ tests/ImageSharp.Benchmarks/Codecs/EncodeBmp.cs
@@ -1,25 +1,20 @@
-﻿// <copyright file="EncodeBmp.cs" company="James Jackson-South">
-// Copyright (c) James Jackson-South and contributors.
+﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
-// </copyright>
 
+using System.Drawing.Imaging;
+using System.IO;
+using BenchmarkDotNet.Attributes;
 using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Tests;
+using SDImage = System.Drawing.Image;
 
-namespace SixLabors.ImageSharp.Benchmarks.Image
+namespace SixLabors.ImageSharp.Benchmarks.Codecs
 {
-    using System.Drawing;
-    using System.Drawing.Imaging;
-    using System.IO;
-
-    using BenchmarkDotNet.Attributes;
-
-    using CoreImage = ImageSharp.Image;
-
+    [Config(typeof(Config.ShortClr))]
     public class EncodeBmp : BenchmarkBase
     {
-        // System.Drawing needs this.
         private Stream bmpStream;
-        private Image bmpDrawing;
+        private SDImage bmpDrawing;
         private Image<Rgba32> bmpCore;
 
         [GlobalSetup]
@@ -27,10 +22,10 @@ namespace SixLabors.ImageSharp.Benchmarks.Image
         {
             if (this.bmpStream == null)
             {
-                this.bmpStream = File.OpenRead("../ImageSharp.Tests/TestImages/Formats/Bmp/Car.bmp");
-                this.bmpCore = CoreImage.Load<Rgba32>(this.bmpStream);
+                this.bmpStream = File.OpenRead(Path.Combine(TestEnvironment.InputImagesDirectoryFullPath, TestImages.Bmp.Car));
+                this.bmpCore = Image.Load<Rgba32>(this.bmpStream);
                 this.bmpStream.Position = 0;
-                this.bmpDrawing = Image.FromStream(this.bmpStream);
+                this.bmpDrawing = SDImage.FromStream(this.bmpStream);
             }
         }
 
@@ -45,7 +40,7 @@ namespace SixLabors.ImageSharp.Benchmarks.Image
         [Benchmark(Baseline = true, Description = "System.Drawing Bmp")]
         public void BmpSystemDrawing()
         {
-            using (MemoryStream memoryStream = new MemoryStream())
+            using (var memoryStream = new MemoryStream())
             {
                 this.bmpDrawing.Save(memoryStream, ImageFormat.Bmp);
             }
@@ -54,7 +49,7 @@ namespace SixLabors.ImageSharp.Benchmarks.Image
         [Benchmark(Description = "ImageSharp Bmp")]
         public void BmpCore()
         {
-            using (MemoryStream memoryStream = new MemoryStream())
+            using (var memoryStream = new MemoryStream())
             {
                 this.bmpCore.SaveAsBmp(memoryStream);
             }
