+++ tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpeg.cs
@@ -1,33 +1,24 @@
-﻿// <copyright file="DecodeJpeg.cs" company="James Jackson-South">
-// Copyright (c) James Jackson-South and contributors.
+﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
-// </copyright>
 
+using System.Drawing;
+using System.IO;
+using BenchmarkDotNet.Attributes;
+using SixLabors.ImageSharp.Formats.Jpeg.GolangPort;
+using SixLabors.ImageSharp.Formats.Jpeg.PdfJsPort;
 using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Tests;
+using CoreSize = SixLabors.Primitives.Size;
+using SDImage = System.Drawing.Image;
 
-namespace SixLabors.ImageSharp.Benchmarks.Image.Jpeg
+namespace SixLabors.ImageSharp.Benchmarks.Codecs.Jpeg
 {
-    using System.Drawing;
-    using System.IO;
-
-    using BenchmarkDotNet.Attributes;
-
-    using SixLabors.ImageSharp.Formats.Jpeg.GolangPort;
-    using SixLabors.ImageSharp.Formats.Jpeg.PdfJsPort;
-    using SixLabors.ImageSharp.Tests;
-
-    using CoreImage = SixLabors.ImageSharp.Image;
-
-    using CoreSize = SixLabors.Primitives.Size;
-    
     [Config(typeof(Config.ShortClr))]
     public class DecodeJpeg : BenchmarkBase
     {
         private byte[] jpegBytes;
 
-        private string TestImageFullPath => Path.Combine(
-            TestEnvironment.InputImagesDirectoryFullPath,
-            this.TestImage);
+        private string TestImageFullPath => Path.Combine(TestEnvironment.InputImagesDirectoryFullPath, this.TestImage);
 
         [Params(TestImages.Jpeg.Baseline.Jpeg420Exif, TestImages.Jpeg.Baseline.Calliphora)]
         public string TestImage { get; set; }
@@ -44,9 +35,9 @@ namespace SixLabors.ImageSharp.Benchmarks.Image.Jpeg
         [Benchmark(Baseline = true, Description = "Decode Jpeg - System.Drawing")]
         public Size JpegSystemDrawing()
         {
-            using (MemoryStream memoryStream = new MemoryStream(this.jpegBytes))
+            using (var memoryStream = new MemoryStream(this.jpegBytes))
             {
-                using (Image image = Image.FromStream(memoryStream))
+                using (var image = SDImage.FromStream(memoryStream))
                 {
                     return image.Size;
                 }
@@ -56,9 +47,9 @@ namespace SixLabors.ImageSharp.Benchmarks.Image.Jpeg
         [Benchmark(Description = "Decode Jpeg - ImageSharp")]
         public CoreSize JpegImageSharpOrig()
         {
-            using (MemoryStream memoryStream = new MemoryStream(this.jpegBytes))
+            using (var memoryStream = new MemoryStream(this.jpegBytes))
             {
-                using (Image<Rgba32> image = CoreImage.Load<Rgba32>(memoryStream, new OrigJpegDecoder()))
+                using (var image = Image.Load<Rgba32>(memoryStream, new OrigJpegDecoder()))
                 {
                     return new CoreSize(image.Width, image.Height);
                 }
@@ -68,9 +59,9 @@ namespace SixLabors.ImageSharp.Benchmarks.Image.Jpeg
         [Benchmark(Description = "Decode Jpeg - ImageSharp PdfJs")]
         public CoreSize JpegImageSharpPdfJs()
         {
-            using (MemoryStream memoryStream = new MemoryStream(this.jpegBytes))
+            using (var memoryStream = new MemoryStream(this.jpegBytes))
             {
-                using (Image<Rgba32> image = CoreImage.Load<Rgba32>(memoryStream, new PdfJsJpegDecoder()))
+                using (var image = Image.Load<Rgba32>(memoryStream, new PdfJsJpegDecoder()))
                 {
                     return new CoreSize(image.Width, image.Height);
                 }
