+++ tests/ImageSharp.Benchmarks/Color/Bulk/PackFromVector4.cs
@@ -1,5 +1,5 @@
 // ReSharper disable InconsistentNaming
-namespace SixLabors.ImageSharp.Benchmarks.Color.Bulk
+namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk
 {
     using System.Numerics;
     using System.Runtime.CompilerServices;
